mdspan: minor comment formatting and rename the header include logic file to Kokkos_MDSpan_Header
